Upgrade services

+++ src/main/java/com/example/photogram/service/impl/PhotoServiceImpl.java
@@ -1,8 +1,14 @@
 package com.example.photogram.service.impl;
 
+import com.example.photogram.exception.NullEntityReferenceException;
 import com.example.photogram.model.Photo;
 import com.example.photogram.repository.PhotoRepository;
 import com.example.photogram.service.PhotoService;
+
+import javax.persistence.EntityNotFoundException;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Optional;
 
 public class PhotoServiceImpl implements PhotoService {
 
@@ -14,16 +20,33 @@ public class PhotoServiceImpl implements PhotoService {
 
     @Override
     public Photo create(Photo photo) {
-        return photoRepository.saveAndFlush(photo);
+        if (photo != null) {
+            return photoRepository.saveAndFlush(photo);
+        }
+        throw new NullEntityReferenceException("Photo cant be 'null'");
     }
 
     @Override
     public void delete(long id) {
-        photoRepository.deleteById(id);
+        Photo photo = readById(id);
+        if (photo != null) {
+            photoRepository.deleteById(id);
+        }
+        throw new EntityNotFoundException("Photo with id: " + id + " not found");
     }
 
     @Override
     public Photo readById(long id) {
-        return photoRepository.findById(id).orElse(null);
+        Optional<Photo> optional = photoRepository.findById(id);
+        if (optional.isPresent()) {
+            return optional.get();
+        }
+        throw new EntityNotFoundException("Photo with id: " + id + " not found");
+    }
+
+    @Override
+    public List<Photo> getPhotosByPostId(long postId) {
+        List<Photo> photos = photoRepository.findPhotosByPostId(postId);
+        return photos.isEmpty() ? new ArrayList<>() : photos;
     }
 }
