+++ src/main/java/com/example/photogram/service/impl/PostServiceImpl.java
@@ -1,10 +1,14 @@
 package com.example.photogram.service.impl;
 
+import com.example.photogram.exception.NullEntityReferenceException;
 import com.example.photogram.model.Post;
 import com.example.photogram.repository.PostRepository;
 import com.example.photogram.service.PostService;
 
+import javax.persistence.EntityNotFoundException;
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 public class PostServiceImpl implements PostService {
 
@@ -16,27 +20,50 @@ public class PostServiceImpl implements PostService {
 
     @Override
     public Post create(Post post) {
-        return postRepository.saveAndFlush(post);
+        if (post != null) {
+            return postRepository.saveAndFlush(post);
+        }
+        throw new NullEntityReferenceException("Post cant be 'null'");
     }
 
     @Override
     public Post update(Post post) {
-        //ToDO: add post check
-        return postRepository.saveAndFlush(post);
+        if (post != null) {
+            Post oldPost = readById(post.getId());
+            if (oldPost != null) {
+                return postRepository.saveAndFlush(post);
+            }
+        }
+        throw new NullEntityReferenceException("Post cant be 'null'");
     }
 
     @Override
     public void delete(long id) {
-        postRepository.deleteById(id);
+        Post post = readById(id);
+        if (post != null) {
+            postRepository.deleteById(id);
+        }
+        throw new EntityNotFoundException("Post with id: " + id + " not found");
     }
 
     @Override
     public Post readById(long id) {
-        return postRepository.findById(id).orElse(null);
+        Optional<Post> optional = postRepository.findById(id);
+        if (optional.isPresent()) {
+            return optional.get();
+        }
+        throw new EntityNotFoundException("Post with id: " + id + " not found");
     }
 
     @Override
     public List<Post> getAll() {
-        return postRepository.findAll();
+        List<Post> posts = postRepository.findAll();
+        return posts.isEmpty() ? new ArrayList<>() : posts;
+    }
+
+    @Override
+    public List<Post> getPostsByUserId(long userId) {
+        List<Post> posts = postRepository.findPostsByUserId(userId);
+        return posts.isEmpty() ? new ArrayList<>() : posts;
     }
 }
